requirements.txt
PyOpenGL==3.1.0
PySDL2==0.9.6
